refactor: add index.js to every screen; rename timetable
