+++ navigation/TabNavigation.jsx
@@ -1,16 +1,12 @@
-import { AntDesign, FontAwesome, MaterialIcons } from '@expo/vector-icons';
+import { AntDesign } from '@expo/vector-icons';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import React from 'react';
-import { Text } from 'react-native';
 
-import ShortTeachPlan from '../screens/shortTeachPlan/shortTeachPlan';
-import Signs from '../screens/signs/Signs';
-import TimeTablePage from '../screens/timeTable/TimeTable';
-import { GLOBAL_STYLES } from '../styles/styles';
+import ShortTeachPlan from '../screens/shortTeachPlan';
+import Signs from '../screens/signs';
+import TimeTablePage from '../screens/timeTable';
 
 const Tab = createBottomTabNavigator();
-
-const renderOptionTitle = (text) => () => <Text style={GLOBAL_STYLES.textIcon}>{text}</Text>;
 
 const TabNavigator = () => (
   <Tab.Navigator
