+++ screens/shortTeachPlan/index.js
@@ -1,0 +1,3 @@
+import ShortTeachPlan from './ShortTeachPlan';
+
+export default ShortTeachPlan
