+++ screens/signs/index.js
@@ -1,0 +1,3 @@
+import Signs from './Signs';
+
+export default Signs;
