+++ screens/timeTable/index.js
@@ -1,0 +1,3 @@
+import TimeTable from './TimeTable';
+
+export default TimeTable;
